CREATE-SPA
Final version of the CREATE-SPA. See readme for specs and installation instructions.

There are a number of changes and improvements I'd implement if given more time:
1. Secure handling of encryption keys
2. HTTPS for communication with server
3. DB backups and load balancing plan
4. Better client side security for protection of sensitive data
5. Validation upon data decryption
6. More comprehensive adherence to HL7® FHIR® standards for communication, data storage, and validation
7. Integration with Ontario government APIs (if available) to further validate patient data

These would each require thorough research to create a list of requirements, and for a real project should be done before work has started so that the project doesn't require significant refactoring as it's being developed.